# backend/services/gateway-service/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Order Fill Service API
  version: 0.1.0
  description: Job API for Excel order-fill and North merge processing.
  license:
    name: Proprietary
    url: https://example.com/proprietary-license
servers:
  - url: http://127.0.0.1:3200
    description: Local app (frontend origin, API via /api)
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Create a session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  type: string
                password:
                  type: string
                  minLength: 10
      responses:
        "200":
          description: Session cookie set, or a two-factor challenge when 2FA is enabled.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/User"
                  - $ref: "#/components/schemas/TwoFactorChallenge"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/login/2fa:
    post:
      operationId: completeTwoFactorLogin
      tags: [auth]
      summary: Complete login with a TOTP or recovery code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                challenge_id:
                  type: string
                code:
                  type: string
      responses:
        "200":
          description: Session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/invite:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Set password from a one-time invite and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  minLength: 10
      responses:
        "200":
          description: Session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Cookie cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout-everywhere:
    post:
      operationId: logoutEverywhere
      tags: [auth]
      summary: End every session for the current user
      responses:
        "204":
          description: All sessions removed and cookie cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/sessions:
    get:
      operationId: listSessions
      tags: [auth]
      summary: List active sessions for the current user
      responses:
        "200":
          description: Active login sessions.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/sessions/{id}/delete:
    post:
      operationId: revokeSession
      tags: [auth]
      summary: End one session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Session removed. Cookie cleared when it was the current session.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password]
              properties:
                current_password:
                  type: string
                password:
                  type: string
                  minLength: 10
      responses:
        "204":
          description: Password updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current session user
      responses:
        "200":
          description: Authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/2fa/setup:
    post:
      operationId: startTwoFactorSetup
      tags: [auth]
      summary: Start app-based two-factor setup
      responses:
        "200":
          description: Pending TOTP secret and QR payload.
          content:
            application/json:
              schema:
                type: object
                required: [secret, otpauth_url, qr_png_base64]
                properties:
                  secret:
                    type: string
                  otpauth_url:
                    type: string
                  qr_png_base64:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/2fa/enable:
    post:
      operationId: enableTwoFactor
      tags: [auth]
      summary: Confirm TOTP setup with a code from the app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Recovery codes shown once.
          content:
            application/json:
              schema:
                type: object
                required: [recovery_codes]
                properties:
                  recovery_codes:
                    type: array
                    items:
                      type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/2fa/disable:
    post:
      operationId: disableTwoFactor
      tags: [auth]
      summary: Disable two-factor authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "204":
          description: Two-factor disabled.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/passkeys/register/begin:
    post:
      operationId: beginPasskeyRegistration
      tags: [auth]
      summary: Start adding a passkey for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Browser options for navigator.credentials.create.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyBegin"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/passkeys/register/finish:
    post:
      operationId: finishPasskeyRegistration
      tags: [auth]
      summary: Finish adding a passkey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, credential]
              properties:
                challenge_id:
                  type: string
                name:
                  type: string
                credential:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Stored passkey public view.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passkey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/passkeys:
    get:
      operationId: listPasskeys
      tags: [auth]
      summary: List passkeys for the current user
      responses:
        "200":
          description: Passkeys without key material.
          content:
            application/json:
              schema:
                type: object
                required: [passkeys]
                properties:
                  passkeys:
                    type: array
                    items:
                      $ref: "#/components/schemas/Passkey"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/passkeys/{id}/delete:
    post:
      operationId: deletePasskey
      tags: [auth]
      summary: Remove a passkey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Passkey removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/passkeys/login/begin:
    post:
      operationId: beginPasskeyLogin
      tags: [auth]
      summary: Start passkey login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
      responses:
        "200":
          description: Browser options for navigator.credentials.get.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyBegin"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/passkeys/login/finish:
    post:
      operationId: finishPasskeyLogin
      tags: [auth]
      summary: Finish passkey login and set the session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, credential]
              properties:
                challenge_id:
                  type: string
                credential:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Session cookie set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Export history visible to the current user
      parameters:
        - name: company_id
          in: query
          description: Platform admin may filter by company. Other roles ignore this.
          schema:
            type: string
      responses:
        "200":
          description: Job list.
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/public/companies/{slug}/login:
    get:
      operationId: getCompanyLogin
      tags: [companies]
      summary: Public login screen metadata for a company slug
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active company login screen metadata. Does not include company id or users.
          content:
            application/json:
              schema:
                type: object
                required: [name, login_slug]
                properties:
                  name:
                    type: string
                  login_slug:
                    type: string
                  has_logo:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/public/companies/{slug}/logo:
    get:
      operationId: getCompanyLogo
      tags: [companies]
      summary: Public company logo for the login screen
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNG, JPEG or WebP image.
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies (platform admin)
      responses:
        "200":
          description: Companies.
          content:
            application/json:
              schema:
                type: object
                required: [companies]
                properties:
                  companies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Company"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company (platform admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, login_slug]
              properties:
                name:
                  type: string
                login_slug:
                  type: string
                  description: Public latin login address, unique among companies.
                matching_mode:
                  type: string
                  enum: [standard, smart]
                  description: Platform admin only. Default standard.
      responses:
        "201":
          description: Company created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/companies/{company_id}/login-slug:
    post:
      operationId: setCompanyLoginSlug
      tags: [companies]
      summary: Set the public latin login address for a company (company admin)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login_slug]
              properties:
                login_slug:
                  type: string
      responses:
        "200":
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/companies/{company_id}/profile:
    post:
      operationId: updateCompany
      tags: [companies]
      summary: Update company name, public login address, and matching mode (platform admin matching_mode only)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, login_slug]
              properties:
                name:
                  type: string
                login_slug:
                  type: string
                matching_mode:
                  type: string
                  enum: [standard, smart]
                  description: Ignored unless the actor is platform_admin.
      responses:
        "200":
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/companies/{company_id}/logo:
    post:
      operationId: setCompanyLogo
      tags: [companies]
      summary: Upload a company logo for the login screen (company admin)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [logo]
              properties:
                logo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/companies/{company_id}/logo/clear:
    post:
      operationId: clearCompanyLogo
      tags: [companies]
      summary: Remove the company login logo (company admin)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Company updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/companies/{company_id}/disable:
    post:
      operationId: disableCompany
      tags: [companies]
      summary: Disable a company (platform admin)
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "204":
          description: Disabled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/companies/{company_id}/users:
    get:
      operationId: listCompanyUsers
      tags: [users]
      summary: List users in a company
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createCompanyUser
      tags: [users]
      summary: Invite a user; raw invite URL is returned once
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, role]
              properties:
                login:
                  type: string
                role:
                  type: string
                  enum: [purchaser, company_admin, company_owner]
      responses:
        "201":
          description: User created with a one-time invite URL.
          content:
            application/json:
              schema:
                type: object
                required: [user, invite_url]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  invite_url:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user_id}/disable:
    post:
      operationId: disableUser
      tags: [users]
      summary: Disable a user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Disabled.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user_id}/reset:
    post:
      operationId: resetUserAccess
      tags: [users]
      summary: Invalidate sessions and issue a new invite
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: New invite URL, shown once.
          content:
            application/json:
              schema:
                type: object
                required: [invite_url]
                properties:
                  invite_url:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/audit:
    get:
      operationId: listAudit
      tags: [audit]
      summary: Recent audit events (platform admin)
      responses:
        "200":
          description: Audit feed.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/status:
    get:
      operationId: listStatus
      tags: [status]
      summary: Dependency health for platform admin
      responses:
        "200":
          description: Component health.
          content:
            application/json:
              schema:
                type: object
                required: [components]
                properties:
                  components:
                    type: array
                    items:
                      $ref: "#/components/schemas/StatusComponent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/order-fill:
    post:
      operationId: createOrderFillJob
      tags: [jobs]
      summary: Create an order-fill workbook processing job
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/OrderFillRequest"
      responses:
        "202":
          description: Job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/jobs/north-merge:
    post:
      operationId: createNorthMergeJob
      tags: [jobs]
      summary: Create a North merge workbook processing job
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/NorthMergeRequest"
      responses:
        "202":
          description: Job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/jobs/{job_id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job status and metadata
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/{job_id}/report:
    get:
      operationId: getJobReport
      tags: [jobs]
      summary: Get report rows for a job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/{job_id}/edits:
    post:
      operationId: submitJobEdits
      tags: [jobs]
      summary: Submit manual edits before finalization
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edits]
              properties:
                edits:
                  type: array
                  items:
                    $ref: "#/components/schemas/ManualEdit"
      responses:
        "202":
          description: Edits accepted and finalization queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/{job_id}/files:
    get:
      operationId: listJobFiles
      tags: [files]
      summary: List generated output files
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Generated files.
          content:
            application/json:
              schema:
                type: object
                required: [files]
                properties:
                  files:
                    type: array
                    items:
                      $ref: "#/components/schemas/OutputFile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: order_fill_session
...
  responses:
...
    Unauthorized:
      description: Session missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found or not visible to the current user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    User:
      type: object
      required: [id, login, role]
      properties:
        id:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [platform_admin, company_owner, company_admin, purchaser]
        company_id:
          type: string
        company_name:
          type: string
        login_slug:
          type: string
        has_logo:
          type: boolean
        two_factor_enabled:
          type: boolean
        has_passkey:
          type: boolean
        last_seen_at:
          type: string
          format: date-time
        disabled_at:
          type: string
          format: date-time
    TwoFactorChallenge:
      type: object
      required: [two_factor_required, challenge_id]
      properties:
        two_factor_required:
          type: boolean
          enum: [true]
        challenge_id:
          type: string
...